% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{buildPackage}
\alias{buildPackage}
\alias{buildAll}
\alias{topN}
\alias{topNCompiled}
\alias{buildUpdatedPackages}
\title{Build a Package}
\usage{
buildPackage(pkg, tgt, debug = FALSE, verbose = FALSE, force = FALSE,
  xvfb = FALSE, suffix = ".1", debver = "1.", plusdfsg = FALSE,
  dryrun = FALSE, compile = FALSE)

buildAll(pkg, tgt, debug = FALSE, verbose = FALSE, force = FALSE,
  xvfb = FALSE)

topN(npkg, date = Sys.Date() - 1, from = 1L)

topNCompiled(npkg, date = Sys.Date() - 1, from = 1L)

buildUpdatedPackages(tgt, debug = FALSE, verbose = FALSE, force = FALSE,
  xvfb = FALSE, bioc = FALSE, dryrun = FALSE)
}
\arguments{
\item{pkg}{character Name of the CRAN or BioConductor package to build}

\item{tgt}{character Name (or version) of the build target distribution, this is restricted
to either \dQuote{20.04}, \dQuote{22.04}, \dQuote{24.04} or \dQuote{26.04} (or their names
\dQuote{focal}, \dQuote{jammy} \dQuote{noble}, or \dQuote{resolute}).}

\item{debug}{logical Optional value to show more debugging output, default is \sQuote{FALSE}}

\item{verbose}{logical Optional value show more verbose progress output, default is \sQuote{FALSE}}

\item{force}{logical Optional value to force package build from source, default is \sQuote{FALSE}}

\item{xvfb}{logical Optional value to build under \code{xvfb-run}, default is \sQuote{FALSE}}

\item{suffix}{character Optional value to override default package version suffix of \sQuote{.1}}

\item{debver}{character Optional value for beginning of Debian build version,
default \sQuote{1.}}

\item{plusdfsg}{logical Optional switch whether \dQuote{+dfsg} gets appended to usptream,
default \sQuote{FALSE}}

\item{dryrun}{logical Optional value to skip actual package build step, default is \sQuote{FALSE}}

\item{compile}{logical Optional value to ensure compilation from source, default is \sQuote{FALSE}}

\item{npkg}{integer Number of packages to build}

\item{date}{Date Relevant date for cranlog download stats}

\item{from}{integer Optional applied as offset to \code{npkg} to shift the selection}

\item{bioc}{logical Optional switch for BioConductor build}
}
\value{
Nothing as the function is invoked for the side effect of building binary packages
}
\description{
Builds a package considering the repository information, dependencies and already
built packages.
}
\details{
The \code{buildPackage} function builds the given package. The \code{buildAll} package applies
to all elements in the supplied vector of packages. The \code{topN} and \code{topNCompiled} helpers
select \sQuote{N} among all (or all compiled) packages. The \code{nDeps} function builds packages
with a given (adjusted) build-dependency count. The \code{buildUpdatedPackages} function finds a 
set of available packages that are not yet built.

Note that this build process is still somewhat tailored to the build setup use by the author and
is not (yet ?) meant to be universally transferable. It should be with a little care and possible
examination of the code. If interested, please get in touch.
}
\seealso{
\code{buildUpdatedPackages}
}
\author{
Dirk Eddelbuettel
}
