% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{buildPackage}
\alias{buildPackage}
\alias{buildAll}
\alias{topN}
\alias{topNCompiled}
\title{Build a Package}
\usage{
buildPackage(pkg, db, repo = c("CRAN", "Bioc"), debug = FALSE,
  verbose = FALSE, force = FALSE, xvfb = FALSE, suffix = ".1")

buildAll(pkg, db, repo = c("CRAN", "Bioc"), debug = FALSE)

topN(npkg, date = Sys.Date() - 1, from = 1L)

topNCompiled(npkg, db, date = Sys.Date() - 1, from = 1L)
}
\arguments{
\item{pkg}{character Name of the CRAN or BioConductor package to build}

\item{db}{data.frame Optional repository information now taken from information loaded at startup}

\item{repo}{character Optional value either \sQuote{CRAN} or \sQuote{Bioc} now also in \code{db} loaded}

\item{debug}{logical Optional value to show more debugging output, default is \sQuote{FALSE}}

\item{verbose}{logical Optional value show more verbose progress output, default is \sQuote{FALSE}}

\item{force}{logical Optional value to force package build from source, default is \sQuote{FALSE}}

\item{xvfb}{logical Optional value to build under \code{xvfb-run}, default is \sQuote{FALSE}}

\item{suffix}{character Optional value to override default of \sQuote{.1} affixed to package version}

\item{npkg}{integer Number of packages to build}

\item{date}{Date Relevant date for cranlog download stats}

\item{from}{integer Optional applied as offset to \code{npkg} to shift the selection}
}
\value{
Nothing as the function is invoked for the side effect of building binary packages
}
\description{
Builds a package considering the repository information, dependencies and already
built packages.
}
\details{
The \code{buildPackage} function builds the given package. The \code{buildAll} package applies
to all elements in the supplied vector of packages. The \code{topN} and \code{topNCompiled} helpers
select \sQuote{N} among all (or all compiled) packages.
}
\author{
Dirk Eddelbuettel
}
